backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
sentence-transformers>=3.0
faiss-cpu>=1.8
anthropic>=0.40
markdown-it-py>=3.0
